test-requirements.txt
pytest-freezegun
pytest-httpserver
pytest-runner
pytest-socket
